+++ app/controllers/project_types_controller.rb
@@ -19,7 +19,7 @@
 # along with this program; if not, write to the Free Software
 # Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.
 
-class ProjectTypesController < ApplicationController #:nodoc:
+class ProjectTypesController < ApplicationController # :nodoc:
   layout 'admin'
   self.main_menu = false
 
@@ -33,7 +33,7 @@ class ProjectTypesController < ApplicationController #:nodoc:
   def index
     @status = params[:status] || 1
 
-    scope = ProjectType.masters_for_table
+    scope = ProjectType.masters_for_table(@status)
     scope = scope.like(params[:name]) if params[:name].present?
 
     @project_count = scope.count