+++ app/overrides/imports/settings.rb
@@ -3,7 +3,7 @@
 #
 # Redmine plugin for xmera called Project Types Plugin.
 #
-# Copyright (C) 2017 - 2022 Liane Hampe <liaham@xmera.de>, xmera.
+# Copyright (C) 2022 Liane Hampe <liaham@xmera.de>, xmera.
 #
 # This program is free software; you can redistribute it and/or
 # modify it under the terms of the GNU General Public License
@@ -19,19 +19,19 @@
 # along with this program; if not, write to the Free Software
 # Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.
 
-# Plugin's routes
-# See: http://guides.rubyonrails.org/routing.html
+# Target is Redmines app/views/imports/settings.html.erb file
+Deface::Override.new(
+  virtual_path: 'imports/settings',
+  name: 'remove-notification-label',
+  remove: 'label[for=import_settings_notifications]',
+  original: 'e6fc4f226646b6cef18b49f9d81aa3ae2c83abfc',
+  namespaced: true
+)
 
-scope '/admin' do
-  resources :project_types do
-    member do
-      post 'archive'
-      post 'unarchive'
-    end
-  end
-end
-
-get '/projects/imports/new',
-    to: 'imports#new',
-    defaults: { type: 'ProjectImport' },
-    as: 'new_projects_import'
+Deface::Override.new(
+  virtual_path: 'imports/settings',
+  name: 'remove-notification-fields',
+  remove: "erb[loud]:contains('import_settings[notifications]')",
+  original: '2394bc1800fbdf1da9bee5d0ba40fdb90c2ead8e',
+  namespaced: true
+)