+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 4.2.0 - 2022-06-29
+
+### Added
+
+* project import based on project types
+
+### Fixed
+
+* (+/-) toggle of project types in custom fields
+* status filter of project types
+
 ## 4.1.2 - 2022-03-05
 
 ### Fixed